Extract BaseDocument from Document in preparation for #26
The role of `BaseDocument` is to be the minimal common denominator for
`Document` and the upcoming `ImportDocument` from #26

`BaseDocument` only has transactions/entries related public methods and
doesn't (yet) broadcast any notification.

The extraction isn't 100% clean, as we can see in the comments I've
added: I couldn't figure out a way to entirely break free of any
schedule spawn related code. That's why we end up propagating
`global_scope` everywhere. It's not very pretty, but it's the best I
could come with.

Tox is still happy.

The next step will be to start merging this branch with `issue_26_base`
from @brownnrl and see if everything happens happily.